refactor: PieceMapper를 controller에서 사용하여, 변환된 정보를 view로 전달
